finished onTargetDelete improve onTargetDelete tests improve notice and assign to developer notifications
git-svn-id: http://svn.labs.intellij.net/webr-dnq/trunk@2435 d982bc04-d64a-0410-bd9d-c7bcd499104c

+++ com/jetbrains/teamsys/dnq/database/EntityOperations.java
@@ -43,8 +43,8 @@ public class EntityOperations {
       }
     }
 
-    // delete itself
-    e.delete();
+    // delete itself; the check is performed, because onDelete constraints could already delete entity 'e'
+    if (!((TransientEntity) e).isRemoved()) e.delete();
   }
 
   private static void executeDestructors(@NotNull Entity e, @NotNull ModelMetaData md) {
